Move csp examples
Fixes #12042
