--- examples/with-strict-csp-hash/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "with-strict-csp-hash",
-  "version": "1.0.0",
-  "scripts": {
-    "dev": "next",
-    "build": "next build",
-    "start": "next start"
-  },
-  "dependencies": {
-    "next": "latest",
-    "react": "^16.7.0",
-    "react-dom": "^16.7.0"
-  },
-  "license": "ISC"
-}
