--- examples/with-strict-csp-hash/pages/index.js
@@ -1,3 +1,0 @@
-export default function Home() {
-  return <div>Hello World</div>
-}
