+++ examples/with-strict-csp/README.md
@@ -1,8 +1,15 @@
-# Strict CSP example
+# Example app with strict CSP generating script hash
 
-If you want to implement a CSP, the most effective way is to follow the [strict CSP](https://csp.withgoogle.com/docs/strict-csp.html) approach. For it to work, we need to generate a nonce on every request.
+This example features how you can set up a strict CSP for your pages including Next.js' inline bootstrap script by hash.
+It defines the CSP by document `meta` tag.
 
-This example uses [Helmet](https://github.com/helmetjs/helmet) to configure the CSP and add the appropriate headers to all server responses. The nonce is generated with [uuid](https://github.com/kelektiv/node-uuid). Then we can pass the nonce to `<Head>` and `<NextScript>` in the custom `<Document>`.
+Note: There are still valid cases for using a nonce in case you need to inline scripts or styles for which calculating a hash is not feasible.
+
+## Deploy your own
+
+Deploy the example using [Vercel](https://vercel.com):
+
+[![Deploy with Vercel](https://vercel.com/button)](https://vercel.com/import/project?template=https://github.com/zeit/next.js/tree/canary/examples/with-strict-csp-hash)
 
 ## How to use
 
@@ -11,9 +18,9 @@ This example uses [Helmet](https://github.com/helmetjs/helmet) to configure the 
 Execute [`create-next-app`](https://github.com/zeit/next.js/tree/canary/packages/create-next-app) with [npm](https://docs.npmjs.com/cli/init) or [Yarn](https://yarnpkg.com/lang/en/docs/cli/create/) to bootstrap the example:
 
 ```bash
-npm init next-app --example with-strict-csp with-strict-csp-app
+npm init next-app --example with-strict-csp-hash with-strict-csp-hash-app
 # or
-yarn create next-app --example with-strict-csp with-strict-csp-app
+yarn create next-app --example with-strict-csp-hash with-strict-csp-hash-app
 ```
 
 ### Download manually
@@ -21,8 +28,8 @@ yarn create next-app --example with-strict-csp with-strict-csp-app
 Download the example:
 
 ```bash
-curl https://codeload.github.com/zeit/next.js/tar.gz/canary | tar -xz --strip=2 next.js-canary/examples/with-strict-csp
-cd with-strict-csp
+curl https://codeload.github.com/zeit/next.js/tar.gz/canary | tar -xz --strip=2 next.js-canary/examples/with-strict-csp-hash
+cd with-strict-csp-hash
 ```
 
 Install it and run:
