--- examples/with-strict-csp/server.js
@@ -1,21 +1,0 @@
-const express = require('express')
-const next = require('next')
-const csp = require('./csp')
-
-const port = parseInt(process.env.PORT, 10) || 3000
-const dev = process.env.NODE_ENV !== 'production'
-const app = next({ dev })
-const handle = app.getRequestHandler()
-
-app.prepare().then(() => {
-  const server = express()
-
-  csp(server)
-
-  server.use(handle)
-
-  server.listen(port, (err) => {
-    if (err) throw err
-    console.log(`> Ready on http://localhost:${port}`)
-  })
-})
